requirements.txt
fastapi
uvicorn
torch
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
python-dotenv
